requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.1
gradio==4.19.2
aiohttp==3.9.3
tenacity==8.2.3
langchain==0.3.19
langchain-huggingface==0.1.2
langchain-chroma==0.2.2
chromadb==0.4.24
sentence-transformers==2.6.1
sqlalchemy==2.0.27
redis==5.0.1
loguru==0.7.2
openai>=1.12.0
